Move adminCheck to OS specific files
